Add files via upload
DNNs for detecting and classifying objects."/objects_data/MobileNetSSD_deploy.caffemodel" and "/objects_data/MobileNetSSD_deploy.prototxt" are models that are going to be used in this code. They are Caffe versions of models called MobileNet-SSD, which use a hybrid of a framework from Google called MobileNet and another framework called Single Shot Detector (SSD) Multibox.